rating.ts and tests fix
